add pytest and create directories
