+++ src/Classes/GeneralTimeEntry.py
