+++ src/main.py
@@ -2,48 +2,10 @@ import sys
 from calendar import monthrange
 from datetime import datetime, date, timezone
 
-from Classes.CATsRow import CATsRow
+from src.Classes.CATsRow import CATsRow
 from ToolApiManager.togglApiManager import TogglApiManager
 from configFileReader import ConfigFileReader
-
-
-def get_formatted_date(prompt: str, default_date: date = date.today()) -> datetime:
-    """
-    Prompts the user for a date and returns a datetime object.
-
-    Args:
-        prompt (str): The prompt message for the user.
-        default_date (date): The default date to use if no input is provided.
-
-    Returns:
-        datetime: A datetime object parsed from user input or the default date.
-    """
-    user_input = input(prompt)
-    if not user_input:
-        return datetime.combine(default_date, datetime.min.time()).replace(tzinfo=timezone.utc)
-
-    date_format = "%d.%m.%Y"
-    try:
-        return datetime.strptime(user_input, date_format)
-    except ValueError:
-        print(f"❌ Invalid date format '{user_input}'. Please use '{date_format}'.")
-        return get_formatted_date(prompt, default_date)
-
-
-def calculate_date_difference(start_dt: datetime, end_dt: datetime) -> int:
-    """
-    Calculates the difference in days between two datetime objects.
-    Args:
-        start_dt (datetime): The start datetime object.
-        end_dt (datetime): The end datetime object.
-    Returns:
-        int: The number of days between the start and end date.
-    """
-    # The original logic is perfect, no need to change
-    return (end_dt - start_dt).days
-
-
-# --- Main Logic ---
+from src.Helpers.dateHelper import get_formatted_date, calculate_date_difference
 
 # Check for command-line arguments and determine the date range
 args = sys.argv[1:]
@@ -81,7 +43,7 @@ else:
     print("❌ Could not determine a valid date range.")
 
 ## read JSON Config file
-fileConfig = ConfigFileReader("config.json")
+fileConfig = ConfigFileReader("../config.json")
 fileConfig.readConfigFile()
 jsonData = fileConfig.fileContent
 
@@ -110,7 +72,7 @@ print("write data to CATs")
 print("write data to myTE")
 
 # TODO adapt data export according to single or multiple day export
-file_name = "cats_entries.txt"
+file_name = "../cats_entries.txt"
 with open(file_name, "w", encoding="utf-8") as file:
     file.write(catsRow.__str__())
 # pickle.dump(catsRow, file)
